refactor(api): 模块化main.go
